Rename functions to libs.
