+++ src/libs/entitiesToMap/index.ts
