+++ src/libs/isTruthy/index.ts
