+++ src/libs/setValuesToMap/__tests__/addSetValueToMap.ts
