+++ src/libs/setValuesToMap/__tests__/removeSetValueToMap.ts
